# api/v2/manufacturer/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FDO Manufacturer Server API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Go FDO Server API Support",
      "url": "https://github.com/fido-device-onboard/go-fdo-server/issues"
    }
  },
  "paths": {
...
    "/api/v1/vouchers": {
      "get": {
        "operationId": "ListVouchers",
        "summary": "Query ownership vouchers (Deprecated)",
        "description": "Returns a list of ownership vouchers matching optional filter criteria.\nResults are ordered by updated_at timestamp in descending order.\nThe CBOR field is omitted from results (use GET /vouchers/{guid} to retrieve full voucher).\n",
        "parameters": [
          {
            "name": "guid",
            "in": "query",
            "description": "Filter by device GUID (hex-encoded)",
            "schema": {
              "type": "string",
              "example": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c",
              "pattern": "^[0-9a-fA-F]{32}$"
            }
          },
          {
            "name": "device_info",
            "in": "query",
            "description": "Filter by device info string",
            "schema": {
              "type": "string",
              "example": "MyDevice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved voucher list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoucherMetadata"
                  }
                },
                "examples": {
                  "emptyList": {
                    "summary": "No vouchers found",
                    "value": []
                  },
                  "voucherList": {
                    "summary": "List of vouchers",
                    "value": [
                      {
                        "guid": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c",
                        "device_info": "Device 1",
                        "created_at": "2024-01-15T10:30:00Z",
                        "updated_at": "2024-01-15T10:30:00Z"
                      },
                      {
                        "guid": "a1b2c3d4e5f6789012345678abcdef01",
                        "device_info": "Device 2",
                        "created_at": "2024-01-14T09:20:00Z",
                        "updated_at": "2024-01-14T09:20:00Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "example": "Invalid GUID: xyz"
              }
            }
          },
          "500": {
            "description": "Internal server error occurred while querying vouchers",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                }
              }
            }
          }
        },
        "tags": [
          "vouchers"
        ]
      },
      "post": {
        "operationId": "InsertVoucher",
        "summary": "Insert ownership voucher(s) (Deprecated)",
        "description": "Inserts one or more ownership vouchers into the database.\nThe request body should contain PEM-encoded ownership voucher(s).\nEach voucher is verified for:\n- Ownership (voucher owner key matches server's configured keys)\n- Cryptographic integrity (signature chain, certificate chains, etc.)\n- Protocol version and header field validity\nDuplicate vouchers (same GUID with same CBOR data) are silently skipped.\nExisting vouchers (same GUID with different CBOR data) are not overwritten and are silently skipped.\nThe operation succeeds (200 OK) even if all vouchers were duplicates/skipped.\n",
        "requestBody": {
          "description": "PEM-encoded ownership voucher(s)",
          "required": true,
          "content": {
            "application/x-pem-file": {
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "example": "-----BEGIN OWNERSHIP VOUCHER-----\n[base64-encoded CBOR data]\n-----END OWNERSHIP VOUCHER-----\n"
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voucher(s) successfully inserted or were duplicates",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          },
          "400": {
            "description": "Invalid voucher data or verification failed",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "examples": {
                  "decodeCborError": {
                    "summary": "Unable to decode CBOR",
                    "value": "Unable to decode cbor"
                  },
                  "invalidVoucher": {
                    "summary": "Voucher verification failed",
                    "value": "Invalid ownership voucher"
                  },
                  "decodePemError": {
                    "summary": "Unable to decode PEM",
                    "value": "Unable to decode PEM content"
                  }
                }
              }
            }
          },
          "413": {
            "description": "Request payload too large (exceeds maximum allowed size).",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "example": "http: request body too large"
              }
            }
          },
          "500": {
            "description": "Internal server error occurred while inserting voucher",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "examples": {
                  "readBodyError": {
                    "summary": "Error reading request body",
                    "value": "Failure to read the request body"
                  },
                  "insertError": {
                    "summary": "Error inserting into database",
                    "value": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "vouchers"
        ]
      }
    },
    "/api/v1/vouchers/{guid}": {
      "get": {
        "operationId": "GetVoucherByGUID",
        "summary": "Get ownership voucher by GUID (Deprecated)",
        "description": "Retrieves a specific ownership voucher by its device GUID.\nReturns the voucher in PEM format with full CBOR data.\n",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "description": "Device GUID (hex-encoded, 32 characters)",
            "required": true,
            "schema": {
              "type": "string",
              "example": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c",
              "pattern": "^[0-9a-fA-F]{32}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the ownership voucher",
            "content": {
              "application/x-pem-file": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "-----BEGIN OWNERSHIP VOUCHER-----\n[base64-encoded CBOR data]\n-----END OWNERSHIP VOUCHER-----\n"
              }
            }
          },
          "400": {
            "description": "Invalid GUID format",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "examples": {
                  "invalidGuid": {
                    "summary": "Invalid GUID",
                    "value": "Invalid GUID"
                  },
                  "invalidFormat": {
                    "summary": "Invalid GUID format",
                    "value": "Invalid GUID format"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Voucher not found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                },
                "example": "Voucher not found"
              }
            }
          },
          "500": {
            "description": "Internal server error occurred",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorString"
                }
              }
            }
          }
        },
        "tags": [
          "vouchers"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VoucherMetadata": {
        "description": "Ownership voucher metadata (excludes CBOR data)",
        "type": "object",
        "properties": {
          "guid": {
            "description": "Device GUID (hex-encoded)",
            "type": "string",
            "example": "3f4e5d6c7b8a9f0e1d2c3b4a5f6e7d8c"
          },
          "device_info": {
            "description": "Device information string",
            "type": "string",
            "example": "Intel NUC Model XYZ"
          },
          "created_at": {
            "description": "Timestamp when voucher was first inserted",
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          },
          "updated_at": {
            "description": "Timestamp when voucher was last updated",
            "type": "string",
            "format": "date-time",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": [
          "guid",
          "device_info",
          "created_at",
          "updated_at"
        ]
      },
      "ErrorString": {
        "description": "A human-readable error message.",
        "type": "string"
      },
...
    },
...
  },
...
}